holy balls! first GUI, what the fuuuu..
The GUI just embeds the CL output in a TextArea, and the Input works
through a TextField. This is ugly, but works now after a lot of Google
and reading javaDocs. at this point, i can start to implement all needed
features for a first version, then give it out for testing.